Remove static props
